refactor(cli): extract login cluster into commands/login.ts (5.0-alpha)
Eleventh cut of the cli.ts god-module split (entangled chain).

- New src/commands/login.ts: cmdLogin (exported) — service-CLI auth with --plan
  preview + retry loop; ensureSecretsBackend (1Password signin preflight) is
  login-only and kept module-private.
- cli.ts imports cmdLogin back; the still-inline cmdSetup calls the imported
  cmdLogin. COMMANDS table unchanged. Prunes now-unused imports (loginServices,
  onepassword check1PasswordStatus/detect1PasswordMode, spawnChild,
  resolveAllAuth).

cli.ts: 3106 -> 2913 lines (-193; -3417 cumulative from 6330, now under 3000).
Typecheck, eslint (0 errors), full suite (2580), prettier all green.
